コミットせな！！！

+++ prepare_utils/transfer_pd_npy.py
@@ -17,10 +17,10 @@ def transfer_pd_npy(timestamp, tmp_dir, dst_dir, wave_n):
     else:
         raise ValueError('wave_n must be 1 or 3')
 
-    """
+
     source_path = os.path.join(wave_dir, timestamp, 'test/ndarray/*')
 
-    for npy_file in tqdm(glob.glob(path)):
+    for npy_file in tqdm(glob.glob(source_path)):
         npy_path = shutil.copy(npy_file, tmp_dir)
     
         # 3波長の場合、コピーしたnpyファイルから1波長のチャンネルを抽出する
@@ -28,25 +28,10 @@ def transfer_pd_npy(timestamp, tmp_dir, dst_dir, wave_n):
             data = np.load(npy_path)
             data_1wave = data[:, :, :, 0:1] #チャンネルの次元を保持するためにスライスを使う
             np.save(npy_path, data_1wave)
-    """
         
     # copy files
     print(dst_dir)
     copy_files = ['configs.json', 'results.json', 'losses.png']
     for file in copy_files:
-        shutil.copy(os.path.join(wave_dir, timestamp, file), dst_dir)
-
-timestamps = ['202310311754','202311012114']
-tmp_dir = ""
-wave_n = 3
-for timestamp in timestamps:
-    dst_dir = f'./aia3wave/{timestamp}'
-    transfer_pd_npy(timestamp, tmp_dir, dst_dir, wave_n)
-    print(f'Finished {timestamp}')
-
-timestamps = ['202306161543','202311031847', '202311031850']
-wave_n = 1
-for timestamp in timestamps:
-    dst_dir = f'./aia211/{timestamp}'
-    transfer_pd_npy(timestamp, tmp_dir, dst_dir, wave_n)
-    print(f'Finished {timestamp}')
+        if os.path.exists(os.path.join(wave_dir, timestamp, file)):
+            shutil.copy(os.path.join(wave_dir, timestamp, file), dst_dir)
